+++ packages/remix-react/components.tsx
@@ -22,6 +22,7 @@ import {
 import type { LinkProps, NavLinkProps } from "react-router-dom";
 import jsesc from "jsesc";
 import type { TrackedPromise } from "@remix-run/deferred";
+import type { Merge } from "type-fest";
 
 import type { AppData, FormEncType, FormMethod } from "./data";
 import type { EntryContext, DeferredLoaderData, AssetsManifest } from "./entry";
@@ -1666,35 +1667,8 @@ type JsonPrimitives =
 
 type NonJsonPrimitives = undefined | Function | symbol;
 
-type SerializeType<T, UseDeferred = false> = UseDeferred extends false
-  ? SerializeTypeImp<T>
-  : T extends [unknown, ...unknown[]]
-  ? {
-      [k in keyof T]: T[k] extends PromiseLike<infer U>
-        ? Promise<SerializeTypeImp<Awaited<U>>>
-        : T[k] extends NonJsonPrimitives
-        ? null
-        : SerializeTypeImp<T[k]>;
-    }
-  : T extends (infer U)[]
-  ? (U extends PromiseLike<infer V>
-      ? Promise<SerializeTypeImp<Awaited<V>>>
-      : U extends NonJsonPrimitives
-      ? null
-      : SerializeTypeImp<U>)[]
-  : T extends object
-  ? {
-      [k in keyof T as T[k] extends PromiseLike<unknown>
-        ? k
-        : T[k] extends NonJsonPrimitives
-        ? never
-        : k]: T[k] extends PromiseLike<infer U>
-        ? Promise<SerializeTypeImp<Awaited<U>>>
-        : SerializeTypeImp<T[k]>;
-    }
-  : SerializeTypeImp<T>;
-
-type SerializeTypeImp<T> = T extends JsonPrimitives
+// TODO: Re-update type to take into account deferred values
+type SerializeType<T, UseDeferred = false> = T extends JsonPrimitives
   ? T
   : T extends NonJsonPrimitives
   ? never
@@ -1711,12 +1685,34 @@ type SerializeTypeImp<T> = T extends JsonPrimitives
   : T extends ReadonlyArray<infer U>
   ? (U extends NonJsonPrimitives ? null : SerializeType<U>)[]
   : T extends object
-  ? {
-      [k in keyof T as T[k] extends NonJsonPrimitives
-        ? never
-        : k]: SerializeType<T[k]>;
-    }
+  ? SerializeObject<UndefinedOptionals<T>>
   : never;
+
+type SerializeObject<T> = {
+  [k in keyof T as T[k] extends NonJsonPrimitives ? never : k]: SerializeType<
+    T[k]
+  >;
+};
+
+/*
+ * For an object T, if it has any properties that are a union with `undefined`,
+ * make those into optional properties instead.
+ *
+ * Example: { a: string | undefined} --> { a?: string}
+ */
+type UndefinedOptionals<T extends object> = Merge<
+  {
+    // Property is not a union with `undefined`, keep as-is
+    [k in keyof T as undefined extends T[k] ? never : k]: T[k];
+  },
+  {
+    // Property _is_ a union with `defined`. Set as optional (via `?`) and remove `undefined` from the union
+    [k in keyof T as undefined extends T[k] ? k : never]?: Exclude<
+      T[k],
+      undefined
+    >;
+  }
+>;
 
 export type UseDataFunctionReturn<T extends DataOrFunction> = T extends (
   ...args: any[]