+++ packages/remix-cloudflare/CHANGELOG.md
@@ -1,5 +1,19 @@
 # `@remix-run/cloudflare`
 
+## 1.6.7
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.6.7`
+
+## 1.6.6
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/server-runtime@1.6.6`
+
 ## 1.6.5
 
 ### Patch Changes